menu_state_custom_game.cpp:split off functions to another file
